Add PyLint
