+++ src/widgetmanager.py
@@ -25,14 +25,18 @@
 # IN THE SOFTWARE.
 # **********
 
-from typing import ItemsView, Optional
+from typing import Dict, ItemsView, Optional, TYPE_CHECKING
 
 from sdl2 import *
 from sdl2.ext import *
 from sdl2.sdlttf import *
 
+from check import CHECK, CheckFailure
 import widget
 import widgettree
+
+if TYPE_CHECKING:  # Avoid circuluar import.
+    from driftwood import Driftwood
 
 
 class WidgetManager:
@@ -45,6 +49,8 @@ class WidgetManager:
         widgets: The dictionary of widgets, sorted by widget id.
         selected: Whether or not this widget is the selected widget. ex. a selected button in a UI menu.
     """
+    driftwood: 'Driftwood'
+    widgets: Dict[int, widget.Widget]
 
     def __init__(self, driftwood):
         self.driftwood = driftwood
@@ -76,7 +82,7 @@ class WidgetManager:
     def __iter__(self) -> ItemsView:
         return self.widgets.items()
 
-    def load(self, filename: str, template_vars: dict = {}) -> Optional[list]:
+    def load(self, filename: str, template_vars: dict) -> Optional[list]:
         """Widget Tree Loader
 
         Loads and inserts a Jinja2 templated JSON descriptor file (a "widget tree") which defines one or more
@@ -88,6 +94,8 @@ class WidgetManager:
 
         Returns a list of widgets created if succeeded, None if failed.
         """
+        if template_vars is None:
+            template_vars = {}
         tree = widgettree.WidgetTree(self, filename, template_vars)
         self.trees.append(tree)
         return tree.widgets
@@ -107,8 +115,8 @@ class WidgetManager:
         # Input Check
         try:
             CHECK(wid, int, _min=0)
-        except CheckFailure as e:
-            self.driftwood.log.msg("ERROR", "Widget", "select", "bad argument", e)
+        except CheckFailure as ex:
+            self.driftwood.log.msg("ERROR", "Widget", "select", "bad argument", ex)
             return False
 
         if wid not in self.widgets:
@@ -209,7 +217,7 @@ class WidgetManager:
     def insert_text(self,
                     contents: str,
                     fontfile: str,
-                    ptsize: int,
+                    pxsize: int,
                     parent: int = None,
                     x: int = None,
                     y: int = None,
@@ -224,7 +232,7 @@ class WidgetManager:
             Args:
                 contents: The contents of the text.
                 fontfile: Filename of the font to render the text with.
-                ptsize: The point size of the text.
+                pxsize: The pixel size of the text.
                 parent: If set, the wid of the parent container.
                 x: The x position of the text on the window. Center if None.
                 y: The y position of the text on the window. Center if None.
@@ -240,7 +248,7 @@ class WidgetManager:
         try:
             CHECK(contents, str)
             CHECK(fontfile, str)
-            CHECK(ptsize, int, _min=1)
+            CHECK(pxsize, int, _min=1)
             if parent is not None:
                 CHECK(parent, int, _min=0)
             else:
@@ -261,12 +269,12 @@ class WidgetManager:
 
         self.__last_wid += 1
 
-        font = self.driftwood.resource.request_font(fontfile, ptsize)
-        if not font:
+        font = self.driftwood.resource.request_font(fontfile, pxsize)
+        if not font or not font.font:
             self.driftwood.log.msg("ERROR", "Widget", "insert_text", "no such font", fontfile)
             return None
 
-        new_widget = widget.TextWidget(self, self.__last_wid, parent, contents, font, ptsize, x, y, width, height,
+        new_widget = widget.TextWidget(self, self.__last_wid, parent, contents, font, pxsize, x, y, width, height,
                                        color)
         self.widgets[self.__last_wid] = new_widget
 
@@ -342,7 +350,7 @@ class WidgetManager:
             self.driftwood.log.msg("ERROR", "Widget", "kill", "bad argument", e)
             return False
 
-        if wid in self.widgets:
+        if wid in self.widgets.items():
             if getattr(self.widgets[wid], "_terminate", None):
                 self.widgets[wid]._terminate()
             del self.widgets[wid]
@@ -374,9 +382,9 @@ class WidgetManager:
     def reset(self) -> bool:
         """Destroy all widgets.
         """
-        for widget in self.widgets:
-            if getattr(self.widgets[widget], "_terminate", None):
-                self.widgets[widget]._terminate()
+        for item in self.widgets.items():
+            if getattr(item, "_terminate", None):
+                item._terminate()
         self.widgets = {}
         self.selected = None
         self.__insert_root_widget()
@@ -390,15 +398,15 @@ class WidgetManager:
                 if (not self.widgets[wid].parent) or self.widgets[self.widgets[wid].parent].active:
                     srcrect = self.widgets[wid].srcrect()
                     dstrect = list(self.widgets[wid].dstrect())
-                    if type(self.widgets[wid]) is widget.ContainerWidget and self.widgets[wid].image:
+                    if isinstance(self.widgets[wid], widget.ContainerWidget) and self.widgets[wid].image:
                         # Draw a container.
                         r = self.driftwood.frame.overlay(self.widgets[wid].image.texture, srcrect, dstrect)
-                        if type(r) is int and r < 0:
+                        if isinstance(r, int) and r < 0:
                             self.driftwood.log.msg("ERROR", "Widget", "_draw_widgets", "SDL", SDL_GetError())
-                    elif type(self.widgets[wid]) is widget.TextWidget and self.widgets[wid].texture:
+                    elif isinstance(self.widgets[wid], widget.TextWidget) and self.widgets[wid].texture:
                         # Draw some text.
                         r = self.driftwood.frame.overlay(self.widgets[wid].texture, srcrect, dstrect)
-                        if type(r) is int and r < 0:
+                        if isinstance(r, int) and r < 0:
                             self.driftwood.log.msg("ERROR", "Widget", "_draw_widgets", "SDL", SDL_GetError())
 
     def __prepare(self) -> None:
@@ -425,8 +433,8 @@ class WidgetManager:
     def _terminate(self) -> None:
         """Cleanup before deletion.
         """
-        for widget in self.widgets:
-            if getattr(self.widgets[widget], "_terminate", None):
-                self.widgets[widget]._terminate()
+        for item in self.widgets.items():
+            if getattr(item, "_terminate", None):
+                item._terminate()
         self.widgets = {}
         TTF_Quit()
